Add document creation and search
Albeit not in the prettiest implementation, we can now search for
files after uploading them.  The search occurs in a standardised way
and the results are easily rendered to the user.

The current approach is fairly straight-forward.  We should provide
wrappers for wrapping around the search solutions which have a
different base path than the standard solutions.  This would create a
new type of resource which we can handle through ember-data but should
clear on demand.